Bump linux and riscv-gnu-toolchain
This now contains the latest toolchain release, along with a our Linux
port applied on top of 4.14-rc8.  The toolchain is in good shape: it's
based on stable releases, with every backported patch having landed
upstraem.  Linux hasn't landed upstream yet, so while there's been some
review it will get rebased (and archived).

This builds and boots on both Spike and Freedom.