Allow users to set the ISA and ABI of the target machine
While this is an SDK for the Freedom U platform, which right now is
RV64GC only, it's also where I've been pointing people who want to build
Linux and a userspace.  This patch allows users to change the ISA and
ABI of the target.

Most of the testing effort has been targeted towards RV64GC, but I think
it's good to let people try other targets (and then hopefully fix them
:)).