Let initramfs automatically mount the rootfs
Root disk image will be written to the sd card at a hardcoded medium.
Therefore initramfs can be notified of this hardcoded address and be
able to mount it.

The root filesystem will be parted at /mnt. The user can only chroot
into that mountpoint. To really boot into the root, /init script should
be written for initramfs rather than be a simple soft link to busybox's
init utility.