Update README with more build instructions
These options roughly include
` Specify hardcoded device and mountpoint for the real root filesystem
in the bbl.bin.
` Specify an offset in units of MB on the target device, where the root
filesystem can be found.
` A make target `make load-sd-card` to format the sd card in one go.
` A make target `make clean-initramfs` that only deletes the BuildRoot
initramfs sysroot build files without touching the kernel build. This
saves a lot of time when recompiling if kernel source file has no
changes.