Allow make to write a qemu filesystem to initramfs
The make target `make ROOT=qemu-elf` will allow the program to package
the qemu-elf as a disk image, which will automatically mounted by
initramfs after boot.

`qemu-elf` comes with a script that relinks the dynamic loader and
shared objects in binaries so that they can run in any riscv system.